"Add function to pluralize strings and return blank for zero count"
Implemented a new function AddSBlank in StringManipulation.ts file. This function takes a string and a count as parameters. It adds an 's' at the end of the input string if the count is more than 1. If the count is zero, it returns a blank string. The function also takes care if the word ends with 's', 'ss', 'z', 'ch', 'sh' or 'x', an 'es' is added. This enhancement was necessary to manage the pluralization of strings based on their counts effectively in the application. Unit tests have also been updated to test the functionality of the new function.