Handle array values in FormDataFromObject
The ObjectConstraint class was not handling array values correctly in the formData append operation. This resulted in incorrect data being appended to the formData object. This patch adds a check to see if the value associated with a key is an array, and if so, appends each item in the array under the same key. This will ensure all data is correctly appended to the formData object.