Add 'nullIfFalsey' option to ObjectConstraint
Updated the ObjectConstraint to allow for a 'nullIfFalsey' option in the fieldConstraint. This provides finer control over how the application should process null or undefined values, specifically if they should be treated as 'false'. Added new test cases to validate this handling. Console log outputs were also adjusted to test this change.