Added EnumValidValue and tests for Enum manipulation
This commit adds seven utility functions to Enums.ts to aid in working with typescript enumerated types. A test for each new function was added to Enums.test.ts file. These utility functions offer more flexibility when working with enumerations, such as obtaining an array of keys and values, finding a key given a value, and checking if a given value is valid within a specific enumeration. These functions were mainly added to offer a more comfortable and efficient way of manipulating and testing enumerated types.