Add AddPrefixToObject utility function
Introduced a new utility function AddPrefixToObject in Functions.ts. This extension adds a prefix to object keys enhancing variable identification, hence making the management of data structures more explicit and readable. This function is demonstrated in consoles.ts where a prefix 'test_' is added to object keys for demonstration purposes.