Add 'CountDecimalDigits' function and tests
A new function 'CountDecimalDigits' has been added to 'Functions.ts', along with corresponding tests in 'Functions.test.ts'. This function accepts a number and counts the number of decimal digits in it. This function is useful to have a precise control of numeric inputs and can help to enhance data validation processes and error handling.