Add IsNumber function and relevant tests
A new function IsNumber is added to help check and ensure that a value is a valid number. This function takes an input as 'value' and returns a boolean indicating whether the provided input is a number or not.

Corresponding tests have also been added to verify the accuracy of this function. Modifications are also made to CountDecimalDigits and DistributeEvenly functions for improved number checking.