Add EqualNumber function and corresponding tests
An EqualNumber function has been included in the Functions.ts file, which allows for determining if two numbers are equal. It also allows specifying the number of decimal places to round to before comparing. Corresponding tests for this new function have been added to Functions.test.ts for verification of working conditions.