Add week number support to DateManager
Additional code was added in the DateFormatAny function in the DateManager.ts file to support display of ISO week numbers. The week number 'w' format has also been included in the listed acceptable formats in the function documentation. This update allows users to include week numbers in their date formats.