Update README.md and add Date Functions documentation
The README.md file has been updated to include information on additional Date helper functions. Specifically, detailed descriptions and examples of the behavior for `DateOnlyNull` were added. In addition, a new file, `Date Functions.md`, has been created under the Documentation directory. It appears to be a placeholder for further descriptions and examples of the different Date functions in the library.