Remove jsLibraryMappings and refactor DateDayOfWeek
The .idea/jsLibraryMappings.xml file has been removed. Additionally, the implementation of DateDayOfWeek in DateManager.ts has been updated with a type definition, changing it from a function to a data type. An array of days of the week, DoWs, has been introduced to spell out days clearly and improve code readability.